webapp/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.2
mysql-connector-python==8.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
flask-cors==4.0.0
python-dateutil==2.8.2
email-validator==2.0.0
random2==1.0.1